requirements.txt
pyrogram
flask
flask[async]
pycryptodome